refactor: move webhook and server into pkg
