+++ cmd/commands/server.go
@@ -19,8 +19,8 @@ package commands
 import (
 	"strings"
 
+	svrcmd "github.com/numaproj/numaflow/pkg/server/cmd"
 	sharedutil "github.com/numaproj/numaflow/pkg/shared/util"
-	svrcmd "github.com/numaproj/numaflow/server/cmd"
 
 	"github.com/spf13/cobra"
 )
