+++ pkg/webhook/validator/pipeline.go
@@ -3,11 +3,12 @@ package validator
 import (
 	"context"
 
+	admissionv1 "k8s.io/api/admission/v1"
+	"k8s.io/client-go/kubernetes"
+
 	dfv1 "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
 	"github.com/numaproj/numaflow/pkg/client/clientset/versioned/typed/numaflow/v1alpha1"
 	pipelinecontroller "github.com/numaproj/numaflow/pkg/reconciler/pipeline"
-	admissionv1 "k8s.io/api/admission/v1"
-	"k8s.io/client-go/kubernetes"
 )
 
 type pipelineValidator struct {
